refactor: refatora a pasta do agente de qa para adequar com a nova arquitetura
