+++ adk/shared/tools/doubt_tool.py
