+++ adk/src/agents/qa_agent/agent.py
@@ -6,8 +6,8 @@ from .subagents.action_planner.agent import agent as action_planner_agent
 from .subagents.code_fix_agent.agent import agent as code_fix_agent
 
 from .subagents.receive_requirements import agent as receber_requisitos_agent
-from .tools.pytest_runner import executar_pytest_tool
-from .tools.doubt_tool import DoubtArtifactGenerator
+from ....shared.tools.pytest_runner import executar_pytest_tool
+from ....shared.tools.doubt_tool import DoubtArtifactGenerator
 
 from .qa_prompt import QA_PROMPT
