--- adk/src/agents/qa_agent/tools/__init__.py
@@ -1,21 +1,0 @@
-from .doubt_artifact import gerar_doubt_artifact
-from .hitl_tool import aguardar_aprovacao_humana
-from .planner_tools import (
-    create_hitl_checkpoint,
-    describe_tools,
-    generate_compliance_report,
-    list_available_tools,
-    plan_validator,
-    register_human_validation,
-)
-
-__all__ = [
-    "aguardar_aprovacao_humana",
-    "create_hitl_checkpoint",
-    "describe_tools",
-    "generate_compliance_report",
-    "gerar_doubt_artifact",
-    "list_available_tools",
-    "plan_validator",
-    "register_human_validation",
-]
